requirements.txt
flask
libvirt-python
xmltodict
pyyaml
evdev
